% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethex.R
\name{gethex}
\alias{gethex}
\title{Extract Hexadecimal Color Codes from Spectral Data}
\usage{
gethex(data, species)
}
\arguments{
\item{data}{Imported images.}

\item{species}{A character string (images name) specifying the name of the species (which can be anything). Use "all" to return hexadecimal color codes for all species.}
}
\value{
If \code{species} is "all", the function returns a dataframe containing a summary of hexadecimal color codes for all species in the dataset. If \code{species} is a specific species name, the function returns the hexadecimal color code for that species. If the specified species does not exist in the dataset, an error message is returned.
}
\description{
This function extracts hexadecimal color codes from imagens (jpg,png). It takes data as input, along with a species name (images name), and returns either a summary of hexadecimal color codes for all species in the dataset or the hexadecimal color code for a specific species.
}
\details{
This function uses the \code{pavo} package to classify spectral data and extract RGB color values. It then converts these RGB values to hexadecimal color codes.
}
\examples{
library(pavo)
library(tidyverse)
data(iris)

# Import images
bf <- pavo::getimg("C:/Users/Diogo Silva/Documents/GitHub/butterfly_spectra/butterspectra/img")

# Extract hexadecimal color code for a specific species
delia <- gethex(bf, species = "delias_aganippe")

# Plot using the extracted color code
ggplot(iris, aes(Sepal.Length, Petal.Width, color = Species)) +
  geom_point() +
  scale_colour_manual(values = delia)

}
\seealso{
\code{\link{classify}}
}
