% File lightSpectra.Rd
\encoding{UTF-8}

\name{LightSpectra}

\title{compute standard light spectra}
\alias{lightSpectra}
\alias{daylightSpectra}
\alias{planckSpectra}
\alias{referenceSpectraTM30}
\alias{illuminantE}

\description{
Two families of standard illuminants that are parameterized by temperature are the
Planckian spectra (black-body spectra), and daylight spectra.
For the daylight spectra, a smoothed version is available.
Illuminant E, a third and trivial spectrum, is also available.
}
\usage{
planckSpectra( temperature, wavelength=300:830, normalize=TRUE, c2=1.4388e-2 )

daylightSpectra( temperature, wavelength=NULL, 
                    components=colorSpec::daylight1964, roundMs=FALSE )
                    
referenceSpectraTM30( temperature, wavelength=380:780, ... )                    

illuminantE( energy=1, wavelength=380:780 )
}

\arguments{
\item{temperature}{a vector of temperatures, in Kelvin}

\item{wavelength}{a vector of wavelengths.
For \code{planckSpectra()} and  \code{illuminantE()} this is required. 
For \code{daylightSpectra()} this is optional. 
The default \code{wavelength=NULL} means to use the wavelengths in \code{components}, 
and otherwise \code{components} is resampled at the given \code{wavelength} vector. }

\item{normalize}{a logical value.  
If \code{TRUE} the Planck spectra are normalized to have value 1 at 560nm.
If \code{FALSE} then the quantity returned is radiant exitance with unit 
\eqn{W * m^{-2} * nm^{-1}}. }
       
\item{c2}{the value of \eqn{hc/k} in Planck's law.
\eqn{h} is the Planck constant; 
\eqn{c} is the speed of light in \eqn{m/sec}; and 
\eqn{k} is the Boltzmann constant.
The default value of 1.4388e-2 \eqn{m*K} was recommended by the CIE in 2005;
in 1986 the CIE recommended \code{c2}=1.438e-2.
If \code{c2='calc'} then \code{c2} is calculated directly from the 3 physical constants,
as recommended by CODATA 2014.
}

\item{components}{a \bold{colorSpec} object with the daylight components \eqn{S_0, S_1}, and \eqn{S_2}.
The default is \code{\link{daylight1964}} and
a smoothed version \code{\link{daylight2013}} is also available.}

\item{roundMs}{a logical value.  The original CIE method for the daylight spectra requires
rounding intermediate coefficients M1 and M2 to 3 decimal places.
This rounding is necessary to reproduce the tabulated values in Table T.1 of the CIE
publication in \bold{References}.}

\item{...}{other args - \code{components} and \code{roundMs} - passed to \code{daylightSpectra}}

\item{energy}{a vector of energy levels}
}
    
\value{
For \code{planckSpectra()}, \code{daylightSpectra()}, and \code{referenceSpectraTM30()}  :\cr
a \bold{colorSpec} object with \code{\link{quantity}} equal to \code{'energy'},
and \code{\link{organization}} equal to \code{'matrix'} or \code{'vector'}.
The \code{specname}s are \code{'PNNNN'} or \code{'DNNNN'}
for \code{planckSpectra()} and \code{daylightSpectra()} respectively.
And for \code{referenceSpectraTM30()} \code{'blended CCT=NNNN'} is used.\cr
The number of spectra in the object is the number of temperatures = \code{length(temperature)}.\cr
  
For \code{illuminantE()} :\cr
A \bold{colorSpec} object with \code{\link{quantity}} equal to \code{'energy'}. \cr
The number of spectra in the object is the number of 
energy levels = \code{length(energy)}.
}


\details{
For \code{planckSpectra()} the valid range of temperatures is 
0 to \code{Inf} (\eqn{\infty}) K, but with exceptions at the endpoints.
For a negative temperature the spectrum is set to all \code{NA}s.
\cr
If \code{temperature=0} and \code{normalize=TRUE}, the spectrum is set to all \code{NA}s.
If \code{temperature=0} and \code{normalize=FALSE}, the spectrum is set to all 0s.
\cr
Conversely, if \code{temperature=Inf} and \code{normalize=FALSE}, 
the spectrum is set to all \code{NA}s.
If \code{temperature=Inf} and \code{normalize=TRUE}, 
the spectrum is set to the pointwise limit \eqn{(560/\lambda)^4} (which appears blue).

For \code{daylightSpectra()} the valid range of temperatures is 4000 to 25000 K.
For a temperature outside this range the spectrum is set to all \code{NA}s.

For \code{referenceSpectraTM30()} the valid range of temperatures is 0 to 25000 K.
The spectrum is computed from the recipe in standard IES TM-30, see \bold{ANSI/IES TM-30-20}.
For a temperature \eqn{\le} 4000K, the reference is the Planck spectrum.
For a temperature \eqn{\ge} 5000K, the reference is the Daylight spectrum.
And for a temperature between 4000K and 5000K, the reference is a weighted blend of
the Planck and Daylight spectra at that temperature.

The equations for \code{daylightSpectra()} and \code{planckSpectra()} are complex 
and can be found in the \bold{References}.

\code{IlluminantE()} is trivial - all constant energy.
}

\references{
Günther Wyszecki and W.S. Stiles.
\bold{Color Science : Concepts and Methods, Quantitative Data and Formulae}.
Second Edition.
Wiley-Interscience. 1982.
page 146.

CIE 15: Technical Report: Colorimetry, 3rd edition.
CIE 15:2004.
Table T.1, pp 30-32, and Note 5 on page 69.

Schanda, Janos.
CIE Colorimetry, in
\emph{Colorimetry: Understanding the CIE System}. 
Wiley Interscience. 2007.
p. 42.

\bold{ANSI/IES TM-30-20}.
\bold{Technical Memorandum: IES Method for Evaluating Light Source Color Rendition}.
\url{https://store.ies.org/product/technical-memorandum-ies-method-for-evaluating-light-source-color-rendition/}
}

\seealso{
\link{daylight}, \code{\link{resample}}, \code{\link{organization}}, \code{\link{quantity}},
\link{materialSpectra}
}

\keyword{light}
