% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspec.R
\name{plotspec}
\alias{plotspec}
\title{plot average spectra}
\usage{
plotspec(rspecdata, legend, ...)
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column containing a wavelength range, named 'wl', and spectra data in remaining columns.}

\item{...}{you can use the arguments of the pavo::aggplot (ex: ylim = c(0, 100), main = "aggplot")}
}
\value{
It is a plot.
}
\description{
Calc and plot the average (central line) and standard deviation (shadow area) of a given spectra, according to a vector of identities. Please, see aggplot function from 'pavo' package.
}
\examples{
#example 1 ## not run
data(leptodactyla)
refletancias <- fixspec(leptodactyla)
plotspec(leptodactyla, ylim = c(0, 80))

}
