% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\docType{data}
\name{transmit}
\alias{transmit}
\title{Transmittance of Pavo cristatus}
\format{
\code{data.frame} with 2 columns and 401 observations
\describe{
  \item{wl}{wavelenght 300-700nm}
  \item{Transmission}{The ocular transmission}
}
}
\source{
Hart (2002)
}
\usage{
transmit
}
\description{
This table contains the ocular transmittance from Hart (2002).
}
\keyword{datasets}
