% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-bluetit.R
\name{vis.bluetit}
\alias{vis.bluetit}
\title{bluetit vismodel}
\usage{
vis.bluetit(rspecdata, background, ...)
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column containing a wavelength range, named 'wl', and spectra data in remaining columns.}

\item{background}{choice one column spectrum data to be the background in RNL model.}
}
\description{
Calculate properties of color (saturation, mean brightness and hue), quantum catch (Qi), chromatic and achromatic contrast, given in JND units. The iluminant is D65. If you want to set the parameters, see vismodel and coldist functions of 'pavo' package.
}
\examples{
#example ## not run
data(flowers)
vis <- vis.bluetit(flowers, background = "Goodenia_heterophylla")

}
