% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-fiddler.R
\name{vis.fiddler}
\alias{vis.fiddler}
\title{Fiddler crab vismodel}
\usage{
vis.fiddler(rspecdata, background)
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column containing a wavelength range, named 'wl', and spectra data in remaining columns.}

\item{background}{choice one column spectrum data to be the background in RNL model.}
}
\description{
Calculate quantum catch (Qi), chromatic and achromatic contrast, given in JND units.
This package uses:
pavo::vismodel parameters: qcatch = "Qi",visual = c(430,520), achromatic = "l",illum = "D65",trans = "ideal")
pavo::coldist parameters: qcatch = NULL, noise = "neural", n = c(1, 1), weber.ref= 'longest', weber = 0.12 (Apis melifera)
}
\examples{
#example ## not run
data(leptodactyla)
vis.fiddler <- vis.fiddler(leptodactyla, background = "X00_background")

}
