% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-human-tri.R
\name{vis.human.tri}
\alias{vis.human.tri}
\title{Fiddler crab vismodel}
\usage{
vis.human.tri(rspecdata, background)
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column containing a wavelength range, named 'wl', and spectra data in remaining columns.}

\item{background}{choice one column spectrum data to be the background in RNL model.}
}
\description{
Calculate quantum catch (Qi), chromatic and achromatic contrast, given in JND units.
The visual peaks based in Pessoa et al. (2014) and weber based on Perini et al. (2009)
Photorreceptors density: (???)
This package uses:
pavo::vismodel parameters: qcatch = "Qi",visual = c(420,530, 565), achromatic = "ml",illum = "D65",trans = "ideal")
pavo::coldist parameters: qcatch = NULL, noise = "neural", n = c(1, 10), weber.ref= 'longest', weber = 0.
}
\examples{
#example ## not run
data(leptodactyla)
vis.human.tri <- vis.human.tri(leptodactyla, background = "X00_background")
}
