% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-podarcis.R
\name{vis.podarcis}
\alias{vis.podarcis}
\title{Podarcis vismodel}
\usage{
vis.podarcis(rspecdata, background, illum = c("D65", "bluesky", "forestshade"))
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column with a wavelength range, named 'wl', and spectra data in the remaining columns.}

\item{background}{Choose one column spectrum data to be the background in the RNL model.}

\item{illum}{You can choose the following illuminants: "D65", "bluesky", and "forestshade".}
}
\description{
Calculate quantum catch (Qi), chromatic and achromatic contrast, given in
JND units. The vismodel parameters of this function are based on Silva et al. (2022).
Parameters: pavo::vismodel(rspecdata, qcatch = "Qi", visual = "podarcis", achromatic = "pd.dc", illum = "D65", trans = transmit, vonkries = FALSE, scale = 1, relative = FALSE)
pavo::coldist(noise = "neural", achro = TRUE, n = c(1, 1.9, 2.7, 2.7), weber.ref = 'longest', weber = 0.1, weber.achro = TRUE)
}
\examples{
# Example ## not run
data(leptodactyla)
leptodactyla <- fixspec(leptodactyla)
vis.podarcis <- vis.podarcis(leptodactyla, background = "X00_background", illum = "D65")

}
