% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspec.R
\name{plotspec}
\alias{plotspec}
\title{plot average spectra}
\usage{
plotspec(rspecdata, ...)
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column containing a wavelength range, named 'wl', and spectra data in remaining columns.}

\item{ylim}{set the limits of y axis.}

\item{main}{plot's title.}
}
\description{
Calc and plot the average (central line) and standard deviation (shadow area) of a given spectra, according to a vector of identities. Please, see aggplot function from 'pavo' package.
}
\examples{
#example 1 ## not run
data(sicalis)
plotspec(sicalis)

#examplo 2 ## not run
plotspec(sicalis, ylim = c(0,50), main = "Sicalis plot")

}
