% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-fiddler.R
\name{vis.fiddler}
\alias{vis.fiddler}
\title{Visual Contrast Analysis Function for Fiddler Crabs}
\usage{
vis.fiddler(rspecdata, background)
}
\arguments{
\item{rspecdata}{A dataset, possibly of rspec class, containing a 'wl' column with wavelength range information and spectral data in the remaining columns.}

\item{background}{Choose one column of spectral data to be used as the background in the RNL model.}
}
\description{
This function calculates the quantum catch (Qi), chromatic contrast, and
achromatic contrast in Just Noticeable Difference (JND) units for fiddler crabs
vision. The visual model parameters are adapted from Silva et al. (2022) and
utilize settings from the 'pavo' package.
}
\examples{
Example Usage:
data(leptodactyla)
vis.fiddler(leptodactyla, background = "X00_background")

}
