% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-fiddler.R
\name{vis.fiddler}
\alias{vis.fiddler}
\title{Fiddler crab vismodel}
\usage{
vis.fiddler(rspecdata, background, ...)
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column containing a wavelength range, named 'wl', and spectra data in remaining columns.}

\item{background}{choice one column spectrum data to be the background in RNL model.}
}
\description{
Calculate properties of color (saturation, mean brightness and hue), quantum catch (Qi), chromatic and achromatic contrast, given in JND units.
}
\examples{
#example ## not run
data(flowers)
vis <- vis.fiddler(flowers, background = "Goodenia_heterophylla")

}
