% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-human-di.R
\name{vis.human.di}
\alias{vis.human.di}
\title{Colorblind human visual modeling}
\usage{
vis.human.di(rspecdata, background)
}
\arguments{
\item{rspecdata}{A data frame, possibly of class rspec, which contains a column containing a wavelength range, named 'wl', and spectra data in remaining columns.}

\item{background}{choice one column spectrum data to be the background in RNL model.}
}
\description{
Calculate quantum catch (Qi), chromatic and achromatic contrast, given in JND units.
The visual peaks based in Pessoa et al. (2014) and weber based on Perini et al. (2009)
Photoreceptors density: (???)
This package uses:
pavo::vismodel(rspecdata, qcatch ="Qi",visual = sens_hum, achromatic = "l", illum = "D65", trans = "ideal", vonkries = FALSE, scale = 1, relative = FALSE)
pavo::coldist(QI_human, qcatch = NULL, noise = "neural", subset = background, achro = TRUE, n = c(1,30), weber.ref = 'l', weber = c(0.08, 0.014), weber.achro = TRUE)
}
\examples{
#example ## not run
data(leptodactyla)
vis.human.di <- vis.human.di(leptodactyla, background = "X00_background")
}
