% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-human-tri.R
\name{vis.human.tri}
\alias{vis.human.tri}
\title{Calculate chromatic and achromatic contrasts for trichromatic human vision}
\usage{
vis.human.tri(rspecdata, background, illum)
}
\arguments{
\item{rspecdata}{A dataframe containing reflectance spectra data. Each row represents a spectrum, and each column represents a wavelength with corresponding reflectance values.}

\item{background}{A character string specifying the ID of the background spectrum used for calculations.}

\item{illum}{A character string specifying the illuminant used for calculations.}
}
\value{
A dataframe containing calculated chromatic and achromatic contrasts along with other relevant information.
}
\description{
This function calculates chromatic and achromatic contrasts for trichromatic human vision based on reflectance spectra data.
}
\examples{
library(pavo)
# Generate example data
rspecdata <- data.frame(Wavelength = seq(300, 700, by = 10), Spectra = runif(41))
background <- "background_spectrum"
illum <- "D65"
# Calculate chromatic and achromatic contrasts
vis.human.tri(rspecdata, background, illum)

}
\seealso{
\code{\link[pavo:sensmodel]{sensmodel}}, \code{\link[pavo:vismodel]{vismodel}}, \code{\link[pavo:coldist]{coldist}}
}
