% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-human-tri.R
\name{vis.human.tri}
\alias{vis.human.tri}
\title{Visual Contrast Analysis Function for Trichromatic Humans}
\usage{
vis.human.tri(rspecdata, background)
}
\arguments{
\item{rspecdata}{A dataset, possibly of rspec class, containing a 'wl' column with wavelength range information and spectral data in the remaining columns.}

\item{background}{Choose one column of spectral data to be used as the background in the RNL model.}
}
\description{
This function calculates the quantum catch (Qi), chromatic contrast (dS), and achromatic
contrast in Just Noticeable Difference (JND) units for trichromatic humans based
on spectral reflectance data. The visual model employs settings
from the 'pavo' package.
--------------------------------
#trichromatic sensibility
sens_hum <- pavo::sensmodel(c(420, 530, 565), range = c(300, 700))
sens_hum <- pavo::as.rspec(sens_hum, lim =c(300, 700))
quantum_catch   <- pavo::vismodel(rspecdata, qcatch ="Qi",visual = sens_hum, achromatic = "ml", illum = "D65", trans = "ideal", vonkries = FALSE, scale = 1, relative = FALSE)
dS  <- pavo::coldist(quantum_catch, qcatch = NULL, noise = "neural", subset = background, achro = TRUE, n = c(1, 6, 13), weber.ref = 'l', weber = c(0.08, 0.02, 0.02), weber.achro = TRUE)
----------------------------------
}
\examples{
# Example Usage:
data(leptodactyla)
vis.human.tri(leptodactyla, background = "X00_background")
}
